Crate proc_macro_plugin

🔬 This is a nightly-only experimental API. (rustc_private)

Proc_Macro

A library for procedural macro writers.

Usage

This crate provides the quote! macro for syntax creation.

The quote! macro uses the crate syntax, so users must declare extern crate syntax; at the crate root. This is a temporary solution until we have better hygiene.

Quasiquotation

The quasiquoter creates output that, when run, constructs the tokenstream specified as input. For example, quote!(5 + 5) will produce a program, that, when run, will construct the TokenStream 5 | + | 5.

Unquoting

Unquoting is done with $, and works by taking the single next ident as the unquoted term. To quote $ itself, use $$.

A simple example is:

fn double(tmp: TokenStream) -> TokenStream {
    quote!($tmp * 2)
}

Large example: Scheme's cond

Below is an example implementation of Scheme's cond.

fn cond(input: TokenStream) -> TokenStream {
    let mut conds = Vec::new();
    let mut input = input.trees().peekable();
    while let Some(tree) = input.next() {
        let mut cond = match tree {
            TokenTree::Delimited(_, ref delimited) => delimited.stream(),
            _ => panic!("Invalid input"),
        };
        let mut trees = cond.trees();
        let test = trees.next();
        let rhs = trees.collect::<TokenStream>();
        if rhs.is_empty() {
            panic!("Invalid macro usage in cond: {}", cond);
        }
        let is_else = match test {
            Some(TokenTree::Token(_, Token::Ident(ident))) if ident.name == "else" => true,
            _ => false,
        };
        conds.push(if is_else || input.peek().is_none() {
            quote!({ $rhs })
        } else {
            let test = test.unwrap();
            quote!(if $test { $rhs } else)
        });
    }

    conds.into_iter().collect()
}

Functions

plugin_registrar 	 [
Experimental
]